Fix update check issue
In some browsers, including Chrome 42, the update check would fail
because the browser would fail to find the .bbSize elements inside
the BSC thread OP element. By checking that the OP element has a
nextSibling before we try to run the update check, we can ensure
that we do not try to do so until the OP element's descendants are
fully loaded.