Fix ACP not being inserted in PM mode
ACP was not inserted in PM mode because there were two fieldsets with
class="s5fieldset", and in PM mode, the one we want is the second one.
Therefore, .querySelector(".s5fieldset") cannot be used to retrieve
said fieldset. As of now, getTAFieldset() is rewritten so that it
starts at the main textarea, then traverses the DOM tree towards the
root and returns the wanted fieldset if it finds it.