Fix Quote Signature button not fitting on line
On tablets and similar devices where the main content column is very
narrow, the Quote Signature button ("Citera sign.") cannot fit on the
same line as the other buttons and will be pushed down to the next line.
This is unacceptable, so I'm adding a large margin-left to it so that it
stays on the same line. The only drawback is that it may intrude into
the section to the left of it, but I consider that a small sacrifice in
this context.