Fix Google button being redundantly inserted
In some browsers, the way the browser retains pages in history may
result in BSC inserting elements even if they already exist when the
user navigates back/forward in history. Specifically, this is the case
on iOS. Therefore, for each element we want to insert, we have to check
that it does not already exist. We already do this in most cases, but
not for the Google button. I'm adding a simple check to make sure it
does not get redundantly inserted.