Enable fixAdHeight on HTTPS
Since we cannot save settings across origins, I have chosen to not run
BSC at all on HTTPS. However, fixAdHeight is such an essential function
that I am making it run on HTTPS anyway. This means that the user cannot
choose whether or not to run it on HTTPS pages. I consider that an
acceptable sacrifice.