Add option to make custom search with DuckDuckGo
Things that are not really finished:

* It doesn't make too much sense to be able to have both Google and DDG
search options. It should be made with radio buttons or a dropdown
menu or something, but I don't know how to do that. You probably
want to fix this before merging.

* It uses the same font for D as the Google G. That's not nice, but I
couldn't figure out how to change createSwecButton properly.

I also added a newline at the end of the file