Hotfix for dark theme flickering bug
I have not managed to reproduce the issue, but some users reported (and
caught on video) that the dark theme would be disabled for a fraction of a
second every time BSC checked if it should be autoset, which was every
fifth second. This meant the user would be flashed 12 times per minute.

This commit should fix that issue.