Make text module available via getGlobal in SASS
It makes sense to be able to use strings from the text module in SASS,
for example as tooltips or content in ::before or ::after
pseudo-elements.

The import statements in the text module had to be changed because they
relied on compilation options not taking effect until compilation of the
actual userscript code.