Do not replace selected when inserting text
Before this commit, text insertion was a destructive action that could
not be undone. For example, if one had selected a chunk of text with the
intention of wrapping it in a spoiler, but accidentally clicked the Edit
button instead, the selected text would be lost.

The behavior of buttons that wrap the selected text in BBCode tags (e.g.
mark, spoiler and embed images) is not affected.