Revert "Use isAuthor instead of isReader"
This reverts commit c033f22d5e92c2807f663937c48d023bfd44a5f9.

Apparently (or, as I figured out four years ago* but had since
forgotten), `isAuthor` doesn't mean "user is author of this post", but
rather "this post is from the thread author". What we're after is
actually `isReader`. Note that it's not in the HTML source, but added
using JS (without XHR, so we don't need to use promises or anything).

* See commit 541522d832de953c8f51665dd13e5f9a90a42de9.