Make mobile site disclaimer detection more sound
While writing the release post about BSC v3.12.0[1], I repeatedly had a
line of actual text automatically removed, because it was erroneously
identified as a mobile site disclaimer by BSC. It looked like this:

    Bla bla [i]Skickades från m.sweclockers.com[/i] bla bla

This commit tweaks the detection of mobile site disclaimers so as to
make it more sound – i.e., so that whenever BSC identifies a disclaimer,
it actually is one. This comes at the potential cost of completeness –
i.e., BSC may fail to remove actual disclaimers should SweClockers
modify their source code in the future. This seems like a good tradeoff
to me, since removing actual content is much worse than failing to
remove disclaimers.

[1] https://www.sweclockers.com/forum/post/18135084