Quote forum post when PM link is used
Before, the PM link next to a forum post only created a blank PM to the
author of the post. As of this commit, the post is quoted in the new PM,
along with any other posts staged for quotation using "Citera flera".

Notably, the generated quote BBCode can't be used verbatim, because
`[quote postid="123456" ...]`, when used in a PM, creates a link to the
PM with ID 123456. We want to link to a forum post, so we must generate
a link manually.

https://www.sweclockers.com/forum/trad/1586687