Update/rewrite improved corrections (#72)
The feature had become broken by recent updates to the host site.

This is a difficult feature to implement and maintain overall. Not only
is it somewhat hard to know where the dialog _should_ be positioned;
browsers also behave differently. The implementation is also different
for LemonIllusion's adaptive layout.

The current implementation is practically perfect in Chrome and just
slightly off in Firefox when transitioning between layout "modes". I
think it's good enough.

Resolves #69.

Related: #70, #71.