Prepare to migrate Nightly users
I've decided to abandon the concept of nightly builds, because it hasn't
worked out very well. In particular, I tend to forget that I should
issue new "stable" releases. The nature and number of users of BSC also
don't really justify the complexity of "nightly" and "stable" builds.

This PR should ensure that current BSC Nightly users are automatically
migrated to "stable".