Patch dark theme toggle (#148)
It was broken by SweClockers' new, responsive design (#124).

The new toggle is simpler – just an icon, similar to the menu lock
toggle – and doesn't need any styles. (It does, however, need a
stylesheet that prepares space for it, so that it doesn't push the
search field to the right when it's inserted.)

I also decided to simplify its tooltip: it's always the same now,
regardless of whether the dark theme is active or not. I figured this
made sense given the new design of the toggle.