Simplify unsupported preference error handling
Instead of showing an obvious error message to the user in the middle of
the preferences menu, we'll fail altogether.

The rationale for the old solution was that we might forget to check
that a new preference was supported by the GUI, so we might merge a
change that would completely break the preferences GUI for users without
even noticing. However, I've found a way to prevent this through tests,
which will be implemented in a separate PR.