Upgrade to TypeScript 4 and Preact 10
Preact 10 is needed for an upcoming PR.

The explicit imports of JSX and the PM links change are necessary for
the code to typecheck with Preact 10 (the latter because render doesn't
return a reference to the rendered Element anymore).

TypeScript 4 and the changes to tsconfig.json are necessary for JSX
fragments (used in the new PM links code) to typecheck.

The ts-jest and eslint upgrades eliminate warnings about TypeScript 4
not being supported.

The no-unused-vars lint rule has to be disabled because otherwise ESLint
thinks Fragment is unused in pm-links.tsx. (It's used implicitly by the
fragment syntax, just like h by the regular JSX syntax.)