Use renderIn instead of render
In Preact 8, elements rendered with `render` were inserted at the end of
the parent if no third argument was given. In Preact X, this is not
guaranteed, and some of our elements were indeed incorrectly placed,
including the quote signature buttons.

`renderIn` is similar to `render`, but has three required parameters.
The virtual node to be rendered is the last parameter instead of the
first one.

More importantly, instead of the optional `replaceNode` from `render`,
`renderIn` takes as argument a function describing where in the parent
the rendered element should be inserted. This function, when given a
placeholder and a parent, must insert the placeholder directly under the
parent (i.e. as a child). The parameter is typed such that the insert
function "knows" the type of the parent element (e.g. HTMLFormElement)
and can access all its properties etc, even those not part of the
Element interface.

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>