Include post when quote signature button is used (#202)
Previously, only the signature was quoted; now, the post and the
signature are quoted together. This puts the signature in context and
makes it easier for other readers to understand why it was quoted.

Notably, the new implementation makes use of AJAX, whereas the old one
was entirely client-side. This is necessary to have the post formatted
properly, not just plain-text. (The source code of signatures can't be
acquired, unfortunately.)