Patch preferences shortcut (#261)
I believe it was broken by the Dark Mode update.

👉 https://www.sweclockers.com/artikel/37572-antligen-halebop-fixar-morkt-tema-pa-sweclockers

The vanilla notification icons use `xlink:href` values like this one:

    /gfx/iconmap.svg?v=20230629#icon_quote

I assume that the purpose of `?v=20230629` is cache invalidation. We
could probably extract it from one of the vanilla icons and use it for
the preferences shortcut as well, but it doesn't feel worth it. If it
ever breaks due to caching, then we can reconsider.

💡 `git show --color-words='pw-notifications|.'`