Add option to redirect home page to archive (#264)
Some users don't like the home page because it doesn't show content in chronological order: articles from several days ago are sometimes bumped, etc. The general consensus, and [explicit recommendation from SweClockers themselves][recommendation], is to use the archive page instead for those who prefer chronological order.

This PR adds an option to redirect the home page to the archive and modify the internal links to the home page to point to the archive instead, as [suggested] by [evil penguin] @ SweClockers. (One might argue that the links need not be modified, since the user will be redirected anyway, but that would be slower, use more data and possibly make the home page appear more popular than it actually is.)

[recommendation]: https://www.sweclockers.com/forum/post/20374336
[suggested]: https://www.sweclockers.com/forum/post/20374374
[evil penguin]: https://www.sweclockers.com/medlem/8224