Lock height of inter-post ads on mobile (#273)
On mobile (or whenever `@media (min-width: 577px)` is false), SweClockers shows ads between forum posts. Unfortunately, their containers don't have a height of their own, so everything below an ad is pushed down when the ad loads. This is very disorienting and annoying, as can be seen by going to [this post] on mobile (or in a narrow window) and scrolling up just a few centimeters or so before the ad loads.

This PR locks the height of inter-post ads to `320px`, which seems to be the actual height of all of them.

[this post]: https://www.sweclockers.com/forum/post/17706168